LayoutTests/platform/chromium/fast/regex/constructor-expected.txt

This test checks use of the regexp constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS re === RegExp(re) is true
PASS re !== new RegExp(re) is true
PASS re === RegExp(re,'i') threw exception TypeError: Cannot supply flags when constructing one RegExp from another.
PASS re !== new RegExp(re,'i') threw exception TypeError: Cannot supply flags when constructing one RegExp from another.
PASS successfullyParsed is true

TEST COMPLETE
